+++ src/pint/models/parameter/__init__.py
@@ -1,3 +1,4 @@
 from pint.models.parameter.parameter import *
 from pint.models.parameter.pair_parameter import pairParameter
 from pint.models.parameter.mask_parameter import maskParameter
+from pint.models.parameter.prefix_parameter import prefixParameter
